+++ apps/app/components/cycles/index.ts
@@ -1,18 +1,15 @@
+export * from "./cycles-list";
 export * from "./active-cycle-details";
-export * from "./cycles-view";
-export * from "./completed-cycles";
+export * from "./active-cycle-stats";
 export * from "./cycles-list-gantt-chart";
-export * from "./all-cycles-board";
-export * from "./all-cycles-list";
+export * from "./cycles-view";
 export * from "./delete-cycle-modal";
 export * from "./form";
 export * from "./gantt-chart";
 export * from "./modal";
 export * from "./select";
 export * from "./sidebar";
-export * from "./single-cycle-list";
 export * from "./single-cycle-card";
-export * from "./empty-cycle";
+export * from "./single-cycle-list";
 export * from "./transfer-issues-modal";
 export * from "./transfer-issues";
-export * from "./active-cycle-stats";
