+++ apps/app/components/cycles/single-cycle-card.tsx
@@ -4,20 +4,17 @@ import Link from "next/link";
 import Image from "next/image";
 import { useRouter } from "next/router";
 
-import { mutate } from "swr";
-
-// services
-import cyclesService from "services/cycles.service";
+// headless ui
+import { Disclosure, Transition } from "@headlessui/react";
 // hooks
 import useToast from "hooks/use-toast";
+// components
+import { SingleProgressStats } from "components/core";
 // ui
 import { CustomMenu, LinearProgressIndicator, Tooltip } from "components/ui";
-import { Disclosure, Transition } from "@headlessui/react";
-import { AssigneesList, Avatar } from "components/ui/avatar";
-import { SingleProgressStats } from "components/core";
-
+import { AssigneesList } from "components/ui/avatar";
 // icons
-import { CalendarDaysIcon, ExclamationCircleIcon } from "@heroicons/react/20/solid";
+import { CalendarDaysIcon } from "@heroicons/react/20/solid";
 import {
   TargetIcon,
   ContrastIcon,
@@ -42,19 +39,13 @@ import {
 import { copyTextToClipboard, truncateText } from "helpers/string.helper";
 // types
 import { ICycle } from "types";
-// fetch-keys
-import {
-  CYCLE_COMPLETE_LIST,
-  CYCLE_CURRENT_LIST,
-  CYCLE_DRAFT_LIST,
-  CYCLE_LIST,
-  CYCLE_UPCOMING_LIST,
-} from "constants/fetch-keys";
 
 type TSingleStatProps = {
   cycle: ICycle;
   handleEditCycle: () => void;
   handleDeleteCycle: () => void;
+  handleAddToFavorites: () => void;
+  handleRemoveFromFavorites: () => void;
   isCompleted?: boolean;
 };
 
@@ -90,6 +81,8 @@ export const SingleCycleCard: React.FC<TSingleStatProps> = ({
   cycle,
   handleEditCycle,
   handleDeleteCycle,
+  handleAddToFavorites,
+  handleRemoveFromFavorites,
   isCompleted = false,
 }) => {
   const router = useRouter();
@@ -100,94 +93,6 @@ export const SingleCycleCard: React.FC<TSingleStatProps> = ({
   const cycleStatus = getDateRangeStatus(cycle.start_date, cycle.end_date);
   const endDate = new Date(cycle.end_date ?? "");
   const startDate = new Date(cycle.start_date ?? "");
-
-  const handleAddToFavorites = () => {
-    if (!workspaceSlug || !projectId || !cycle) return;
-
-    const fetchKey =
-      cycleStatus === "current"
-        ? CYCLE_CURRENT_LIST(projectId as string)
-        : cycleStatus === "upcoming"
-        ? CYCLE_UPCOMING_LIST(projectId as string)
-        : cycleStatus === "completed"
-        ? CYCLE_COMPLETE_LIST(projectId as string)
-        : CYCLE_DRAFT_LIST(projectId as string);
-
-    mutate<ICycle[]>(
-      fetchKey,
-      (prevData) =>
-        (prevData ?? []).map((c) => ({
-          ...c,
-          is_favorite: c.id === cycle.id ? true : c.is_favorite,
-        })),
-      false
-    );
-
-    mutate(
-      CYCLE_LIST(projectId as string),
-      (prevData: any) =>
-        (prevData ?? []).map((c: any) => ({
-          ...c,
-          is_favorite: c.id === cycle.id ? true : c.is_favorite,
-        })),
-      false
-    );
-
-    cyclesService
-      .addCycleToFavorites(workspaceSlug as string, projectId as string, {
-        cycle: cycle.id,
-      })
-      .catch(() => {
-        setToastAlert({
-          type: "error",
-          title: "Error!",
-          message: "Couldn't add the cycle to favorites. Please try again.",
-        });
-      });
-  };
-
-  const handleRemoveFromFavorites = () => {
-    if (!workspaceSlug || !projectId || !cycle) return;
-
-    const fetchKey =
-      cycleStatus === "current"
-        ? CYCLE_CURRENT_LIST(projectId as string)
-        : cycleStatus === "upcoming"
-        ? CYCLE_UPCOMING_LIST(projectId as string)
-        : cycleStatus === "completed"
-        ? CYCLE_COMPLETE_LIST(projectId as string)
-        : CYCLE_DRAFT_LIST(projectId as string);
-
-    mutate<ICycle[]>(
-      fetchKey,
-      (prevData) =>
-        (prevData ?? []).map((c) => ({
-          ...c,
-          is_favorite: c.id === cycle.id ? false : c.is_favorite,
-        })),
-      false
-    );
-
-    mutate(
-      CYCLE_LIST(projectId as string),
-      (prevData: any) =>
-        (prevData ?? []).map((c: any) => ({
-          ...c,
-          is_favorite: c.id === cycle.id ? false : c.is_favorite,
-        })),
-      false
-    );
-
-    cyclesService
-      .removeCycleFromFavorites(workspaceSlug as string, projectId as string, cycle.id)
-      .catch(() => {
-        setToastAlert({
-          type: "error",
-          title: "Error!",
-          message: "Couldn't remove the cycle from favorites. Please try again.",
-        });
-      });
-  };
 
   const handleCopyText = () => {
     const originURL =
